requirements.txt
pandas
numpy
matplotlib
opencv-python
pillow